Agency Dashboard: update site-table and site-table-row component unit tests
Update the unit tests to reflect the latest changes in the components
