+++ client/jetpack-cloud/sections/agency-dashboard/sites-overview/site-error-content.tsx
@@ -12,11 +12,14 @@ export default function SiteErrorContent( { siteUrl }: { siteUrl: string } ) {
 	};
 
 	return (
-		<div className="sites-overview__error-container">
+		<div className="sites-overview__error-container" data-testid="site-error-container">
 			<span className="sites-overview__error-icon">
 				<Gridicon size={ 18 } icon="notice-outline" />
 			</span>
-			<span className="sites-overview__error-message sites-overview__error-message-large-screen">
+			<span
+				className="sites-overview__error-message sites-overview__error-message-large-screen"
+				data-testid="site-error-message"
+			>
 				{ translate( 'Jetpack is unable to connect to %(siteUrl)s', {
 					args: {
 						siteUrl,
@@ -30,6 +33,7 @@ export default function SiteErrorContent( { siteUrl }: { siteUrl: string } ) {
 				onClick={ handleClickFixNow }
 				className="sites-overview__error-message-link"
 				href={ `/settings/disconnect-site/${ siteUrl }?type=down` }
+				data-testid="site-error-message-link"
 			>
 				{ translate( 'Fix now' ) }
 			</a>
